+++ app/src/main/java/com/example/ronjc/tiptracker/model/Income.java
@@ -21,12 +21,12 @@ public class Income implements Serializable{
     private long date;
     private String category;
     private String userID = "";
-    private long longitude;
-    private long latitude;
+    private double longitude;
+    private double latitude;
 
     public Income() {}
 
-    public Income(String id, String name, double amount, long date, String category, String userID, long longitude, long latitude) {
+    public Income(String id, String name, double amount, long date, String category, String userID, double longitude, double latitude) {
         this.id = id;
         this.name = name;
         BigDecimal bigDecimal = new BigDecimal(amount);
@@ -65,11 +65,11 @@ public class Income implements Serializable{
         return userID;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 }
